2nd_Place/requirements.txt
matplotlib==3.3.1
scipy==1.4.1
torchvision==0.5.0
efficientnet_pytorch==0.7.0
tqdm==4.43.0
xgboost==1.0.2
pandas==1.1.2
hyperopt==0.2.3
torch==1.8.0
torchvision=0.9.0
numpy==1.19.2
Pillow==8.1.0
scikit_learn==0.24.1
